finished factoring embeddings structure
